backend/flask/pyproject.toml
[tool.poetry]
name = "agent-swarm-backend"
version = "0.1.0"
description = ""
authors = ["BrianP8701 <brianp8701@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.1"
flask-cors = "^4.0.0"
flask-jwt-extended = "^4.6.0"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.1"
pydantic = "^2.6.1"
openai = "^1.12.0"
pymongo = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
